--- a/examples/multigpu/graphbolt/node_classification.py
+++ b/examples/multigpu/graphbolt/node_classification.py
@@ -36,6 +36,7 @@ main
       │
       └───> Test set evaluation
 """
+
 import argparse
 import os
 import time