--- a/tests/python/pytorch/graphbolt/test_item_sampler.py
+++ b/tests/python/pytorch/graphbolt/test_item_sampler.py
@@ -1149,7 +1149,9 @@ def test_RangeCalculation(params):
     assert key == answer
 
 
-@unittest.skipIf(os.getenv("DGLTESTDEV", "cpu") != "cpu", reason="GPU not required.")
+@unittest.skipIf(
+    os.getenv("DGLTESTDEV", "cpu") != "cpu", reason="GPU not required."
+)
 @pytest.mark.parametrize("num_ids", [24, 30, 32, 34, 36])
 @pytest.mark.parametrize("num_workers", [0, 2])
 @pytest.mark.parametrize("drop_last", [False, True])
@@ -1250,7 +1252,9 @@ def distributed_item_sampler_subprocess4(
         dist.destroy_process_group()
 
 
-@unittest.skipIf(os.getenv("DGLTESTDEV", "cpu") != "cpu", reason="GPU not required.")
+@unittest.skipIf(
+    os.getenv("DGLTESTDEV", "cpu") != "cpu", reason="GPU not required."
+)
 @pytest.mark.parametrize("num_ids", [24, 30, 32, 34, 36])
 @pytest.mark.parametrize("num_workers", [0, 2])
 @pytest.mark.parametrize("drop_last", [False, True])