@Rhett-Ying 2024-05-06
> num_ids: 36, num_workers: 2
`num_ids` is the total number of `ItemSet` or `ItemSetDict`? If yes, it's too small and not persuasive.